Preparation for proper mealdto to meal conversion

+++ backend/Presentation/Controllers/MealController.cs
@@ -25,6 +25,8 @@ namespace backend.Presentation.Controllers
         [HttpGet]
         public ActionResult<IEnumerable<MealDto>> Get()
         {
+            //TODO: get meals vom Mealservice, get (measuredIngredients) + ingredients
+            //von deren Service und dann hier zusammenfügen
             List<MealDto> tmplist = new();
             foreach(Meal m in _mealService.GetMeals())
             {
@@ -70,6 +72,7 @@ namespace backend.Presentation.Controllers
         [HttpDelete("{id}")]
         public ActionResult<MealDto> Delete(int id)
         {
+            if (id < 0) return BadRequest("Meal Id cannot be negative");
             try
             {
                 var result = _mealService.RemoveMealById(id);
